move alert factory helper to common component
