+++ Wire-iOS/Sources/UserInterface/Helpers/UIViewController+Alert.swift
@@ -19,41 +19,6 @@
 import Foundation
 import UIKit
 
-typealias AlertActionHandler = (UIAlertAction) -> Void
-
-extension UIAlertController {
-
-    /// Create an alert with a OK button
-    ///
-    /// - Parameters:
-    ///   - title: optional title of the alert
-    ///   - message: message of the alert
-    ///   - okActionHandler: a nullable closure for the OK button
-    /// - Returns: the alert presented
-    static func alertWithOKButton(title: String? = nil,
-                                  message: String,
-                                  okActionHandler: AlertActionHandler? = nil) -> UIAlertController {
-        let alert = UIAlertController(title: title,
-                                      message: message,
-                                      preferredStyle: .alert)
-
-        let okAction =  UIAlertAction.ok(style: .cancel, handler: okActionHandler)
-        alert.addAction(okAction)
-
-        return alert
-    }
-
-    convenience init(title: String? = nil,
-                     message: String,
-                     alertAction: UIAlertAction) {
-        self.init(title: title,
-                  message: message,
-                  preferredStyle: .alert)
-        addAction(alertAction)
-    }
-
-}
-
 extension UIViewController {
 
     /// Present an alert with a OK button
