+++ Wire-iOS/Sources/UserInterface/Helpers/UIViewController+ErrorAlert.swift
@@ -18,6 +18,7 @@
 
 import UIKit
 import WireSyncEngine
+import WireCommonComponents
 
 extension UIViewController {
